requirements.txt
pymongo>=4.0.0
sqlalchemy>=2.0.0
click>=8.0.0
pydantic>=2.0.0
rich>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
typer>=0.9.0 